- changed Duke actor setup to exclusively work with scripted classes.
Even though the code is still native the classes cannot be. Duke requires the ability to change class types at run time as a means of state maintenance and this can only be done with the PClass descriptor, not with native VTables.

- ported Duke's crane to proper actor class management.

Code is still native to make debugging easier in the transition phase.

# Conflicts:
#	source/games/duke/src/actors.cpp
#	source/games/duke/src/actors_d.cpp
#	source/games/duke/src/actors_r.cpp
#	source/games/duke/src/duke3d.h
#	source/games/duke/src/game.cpp
#	source/games/duke/src/premap.cpp
#	source/games/duke/src/spawn_r.cpp
#	source/games/duke/src/types.h
#	wadsrc/static/zscript/games/duke/actors/DukeActor.zs